Move scope computation to a new file.
