+++ auth-server/auth_server/users/models.py
@@ -5,8 +5,6 @@ from typing import Annotated, Literal, Sequence, TypedDict
 
 from pydantic import BaseModel, Field, SecretStr
 
-from server_utils.auth.scopes import Scope
-
 
 # leave this outside of the db. this will not change.
 class AccountType(StrEnum):
@@ -16,35 +14,6 @@ class AccountType(StrEnum):
     USER = "user"
     AUDITOR = "auditor"
     SERVICE = "service"
-
-
-# move this to db if we need to support updating scopes.
-ACCOUNT_TYPE_TO_SCOPES: dict[AccountType, set[Scope]] = {
-    AccountType.ADMIN: set(Scope),  # all scopes
-    AccountType.SERVICE: set(Scope),  # all scopes
-    AccountType.USER: {
-        Scope.RESTART_WRITE,
-        Scope.ROBOT_CONTROL_WRITE,
-        Scope.ROBOT_SETTINGS_WRITE,
-        # todo(mm, 2026-03-17): Updates should be togglable to admin-only by an auth setting.
-        Scope.UPDATES_WRITE,
-        Scope.USERS_READ_SELF,
-        Scope.USERS_WRITE_SELF,
-        # todo(mm, 2026-03-17): Protocol uploads should be togglable to admin-only by an auth setting.
-        Scope.PROTOCOLS_WRITE,
-        Scope.AUDIT_LOG_WRITE,
-    },
-    # Auditors should have read-only access to everything. Our read-only endpoints are
-    # mostly accessible without authentication, but there are some exceptions. This
-    # just needs to have the scopes to cover those exceptions.
-    AccountType.AUDITOR: {Scope.USERS_READ_OTHERS},
-}
-
-# Scopes granted while resetPassword is true, before the user chooses a new password.
-RESET_PASSWORD_SCOPES: set[Scope] = {
-    Scope.USERS_READ_SELF,
-    Scope.USERS_WRITE_SELF,
-}
 
 
 class UserCreate(BaseModel):
